+++ integration/spark/src/main/common/java/io/openlineage/spark/agent/facets/builder/EnvironmentFacetBuilder.java
@@ -1,0 +1,2 @@
+package io.openlineage.spark.agent.facets.builder;public class EnvironmentFacetBuilder {
+}