+++ integration/spark/src/main/common/java/io/openlineage/spark/api/OpenLineageContext.java
@@ -6,6 +6,7 @@ import io.openlineage.client.OpenLineage.OutputDataset;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 import lombok.Builder;
 import lombok.Builder.Default;
 import lombok.NonNull;
@@ -28,10 +29,15 @@ import scala.PartialFunction;
  * lists. As {@link QueryPlanVisitor}s require a reference to the {@link OpenLineageContext}, the
  * lists will always be added to after the {@link QueryPlanVisitor}s are constructed. Thus, copies
  * should never be made of the lists, as it should be assumed such copies will be incomplete.
+ *
+ * <p>This API is evolving and may change in future releases
+ *
+ * @apiNote This interface is evolving and may change in future releases
  */
 @Value
 @Builder
 public class OpenLineageContext {
+  UUID runUuid = UUID.randomUUID();
 
   /**
    * Optional {@link SparkSession} instance when an application is using a Spark SQL configuration