+++ proxy/src/main/java/io/openlineage/proxy/ProxyStreamConfig.java
@@ -1,0 +1,27 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.openlineage.proxy;
+
+import com.fasterxml.jackson.annotation.JsonSubTypes;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import io.openlineage.proxy.api.models.ConsoleConfig;
+import io.openlineage.proxy.api.models.KafkaConfig;
+
+@JsonSubTypes({
+  @JsonSubTypes.Type(value = ConsoleConfig.class, name = "Console"),
+  @JsonSubTypes.Type(value = KafkaConfig.class, name = "Kafka")
+})
+@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "type")
+public interface ProxyStreamConfig {}